+++ docs/changelog.md
@@ -17,10 +17,14 @@ Release is mostly focused on tidying up the docs and code base and on resource c
 ### Features
 * New `--raft.logdb` flag (defaults to Tan).
 * Added a raft cluster snapshot mode option `--raft.snapshot-recovery-type` which defaults to new `checkpoint` mode.
+* Added a flag for configuring shared table cache size `--storage.table-cache-size` which defaults to `1024`.
 
 ### Improvements
 * Compressed Log replication messages to lower API bandwidth usage.
 * Bump to Go 1.20.
+* Pebble instances table cache is now shared.
+* Single key lookups now utilise bloom filters optimizing the "key not present" case.
+* Added compaction metrics of pebble instances.
 
 ### Bugfixes
 * Removing kafka package leftovers.